Major refactor to leverage SDK code
