+++ weave/flow/model.py
@@ -8,9 +8,9 @@ from typing import Any, Callable, Optional, Union
 from rich import print
 
 from weave.flow.obj import Object
+from weave.trace.call import async_call_op
 from weave.trace.isinstance import weave_isinstance
 from weave.trace.op import Op, OpCallError, as_op, is_op
-from weave.trace.op_caller import async_call_op
 from weave.trace.weave_client import Call
 
 INFER_METHOD_NAMES = {"predict", "infer", "forward", "invoke"}
